requirements.txt
aws-psycopg2==1.2.1
cassandra-driver==3.24.0
click==7.1.2
geomet==0.2.1.post1
six==1.15.0
